Integrate prisma-fmt based autocompletion
The new method is textDocumentCompletion(). It takes a CompletionParams
request as specified by the language server protocol, and returns a
CompletionList, so it's trivial to integrate.

It currently completes only two kinds of argument attributes: `map` on
`@default()` attributes, based on the datasource, and referential
actions (onDelete and onUpdate), also based on the datasource.

This PR replaces the existing referentialActions completions with the
prisma-fmt based ones.