tech(ls): test reorg (#1749)
Shuffle tests around so that instead of having one multifile dir with completions tests etc, we group them together so that multifile tests are with their respective functions, i.e. all completions tests, rename tests, linting, etc are each together. Moved fixtures to be next to the tests instead of in an entirely separate dir.

---------

Co-authored-by: Serhii Tatarintsev <tatarintsev@prisma.io>
Co-authored-by: Joël Galeran <Jolg42@users.noreply.github.com>