fix: fix config loading crash (#1933)
[TML-1628](https://linear.app/prisma-company/issue/TML-1628/fix-vscode-extension-crash-when-loading-config)

Needed to refactor the code a bit to actually make sure
`PrismaSchema.load` is used by the tests.

Note that if we want to fix it for the users of Prisma 6 we need to
backport this to Prisma 6 language server and release a patch. We don't
really need to release a patch of the Prisma 6 VSCode extension, but we
do need to publish a patched language server for Prisma 6 to NPM and use
it from the Prisma 7 extension.