ci: allow manual dispatch to publish the LS (#1936)
Changes the workflow slightly so that we can trigger it manually to
release the language server only.

The future plan for managing the Prisma 6 support in the extension is to
advise users to always use the latest version of the extension (which
will bundle the Prisma 6 language server). We will not release Prisma 6
extension patches, we will only patch the language server. Therefore we
only need to publish the language server and we'll do that by running
the `3_LS_tests_publish.yml` workflow on a prisma 6 branch. When a new
language server for Prisma 6 is released we need to make sure its
dependency is bumped on our main branch and gets bundled with the latest
extension.