fix: delete expired/failed refresh tokens (#1941)
While testing the VSCode extension I noticed that for me the refresh
action was not working properly. This was due to old expired refresh
tokens on my machine. As the extension failed to refresh the token but
never got rid of the expired token on failure it basically hang.

This PR fixes this by better handling the error case of a token refresh
and deleting the stored credentials if the refresh token is
expired/invalid.

To test this locally you can modify your
`~/Library/Preferences/prisma-platform/auth.json` file to contain an
invalid refresh token. E.g. make it look like this:

```
{
  "tokens": [
    {
      "refreshToken": "pseudo_expired_refresh_token",
      "token": "pseudo_expired_token",
      "workspaceId": "rz3339rmxkxa9qos9mgz71fe"
    }
  ]
}
```